Remodify pretrain

+++ src/train_flow/pre_processing.py
@@ -2,7 +2,7 @@
 # Author : Junhao Wei
 # @file : pre_processing.py
 # @Time : 2024/7/14 17:30
-# Interpretation: Flow of pre-processing.
+# Interpretation: 1. Flow of pre-processing.
 import os.path
 
 import numpy as np
@@ -54,6 +54,7 @@ def get_processed_data(input_df: str, use_features: list, one_hot=False, cat_fea
         feats, labels = data_combined.drop(columns=['InjurySeverity']), data_combined['InjurySeverity']
     else:
         feats, labels = df_features.drop(columns=['InjurySeverity']), df_features['InjurySeverity']
+    feats = feats.astype('int')
 
     x_train, x_test, y_train, y_test = train_test_split(feats, labels, stratify=labels,
                                                         shuffle=True, test_size=0.20,
@@ -62,7 +63,7 @@ def get_processed_data(input_df: str, use_features: list, one_hot=False, cat_fea
     assert round(len(y_train[y_train == 2]) / len(y_train), 3) == round(len(y_test[y_test == 2]) / len(y_test), 3), \
         'The proportion of severe samples in the training and testing sets is different！'
 
-    # Save train test dataset
+    # TODO: re-open Save train test dataset after one-hot only pre-train finished!
     train_test_data_dict = {'x_train': x_train, 'y_train': y_train, 'x_test': x_test, 'y_test': y_test}
     train_test_save_dir = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(__file__))),
                                        'data', 'Combined', one_hot_str)
